Extract slot serialisation logic and simplify SidebarBoards
